finally fixed nasty bug when reading frames of unknown size
The bug had multiple stages.

The first one was my assumption that LZ4F_getFrameInfo returned an error
the first time it was called.
But because I did not update bytes_processed in decompress_single_frame()
the function was called twice, thinking there was a second frame coming.

Then there were some logic errors and in the end it was a question of
statement ordering.

After the last chunk of a frame was decompressed the code would break out
of the decompression loop. But the code updating bytes_processed was
behind that break and so was not called.

But now everything looks fine. I also added some comments for better
understanding.